+++ src/App.tsx
@@ -1,0 +1,22 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, Text, View, Button } from 'react-native';
+
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>이상헌 첫 리액트 네이티브!!!!!!!</Text>
+      <StatusBar style="auto" />
+      <Button title="button" onPress={() => alert('Click')} />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
